test(runtime): split the log stream suite into coherent modules
test_log_stream.py crossed the repository's 700-line review bar at 814
lines. The coverage now lives in three focused modules, every assertion
preserved: test_log_stream.py keeps the reader archiving and resilience
cases, test_log_stream_markers.py holds the parse/emit/boundary-scanning
suite (including the chunked-stream helper), and
test_log_stream_targets.py covers allowlist, sanitization, containment,
resolver, and step-event behavior.
